ftsplitleft

+++ libft/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: zbin-md- <zbin-md-@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/05 12:33:34 by zbin-md-          #+#    #+#             */
-/*   Updated: 2024/06/05 12:41:39 by zbin-md-         ###   ########.fr       */
+/*   Updated: 2024/06/26 20:54:35 by zbin-md-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include "libft.h"
+
+size_t	ft_strlen(char *str)
 {
 	int	strcount;
